requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pandas==2.2.0
numpy==1.26.4
polars==0.20.7
pyarrow==15.0.0
python-multipart==0.0.9
sqlalchemy==2.0.27
pyodbc==5.0.1
openpyxl==3.1.2
pyyaml==6.0.1
scikit-learn>=1.3.0
aiosqlite>=0.19.0  # Async SQLite support
psycopg2-binary>=2.9.0  # PostgreSQL driver
ibm-db-sa>=0.3.9  # IBM Db2 SQLAlchemy adapter
pymysql>=1.1.0  # MySQL driver
cx-oracle>=8.3.0  # Oracle database driver
sentence-transformers==2.5.1
torch>=2.2.0
transformers>=4.38.0
ctransformers>=0.2.27
great-expectations==0.18.22